requirements.txt
appdirs==1.4.3
netifaces==0.10.5
packaging==16.8
paho-mqtt==1.2
psutil==5.2.0
pyparsing==2.2.0
pyspectator==1.0.8
six==1.10.0
